Create chart query hook, move all logic into that and reuse at vitals one
