+++ packages/web/app/api/queries/useVitals.js
@@ -1,72 +1,8 @@
 import { useQuery } from '@tanstack/react-query';
-import { VISIBILITY_STATUSES, VITALS_DATA_ELEMENT_IDS } from '@tamanu/constants';
+import { VITALS_DATA_ELEMENT_IDS } from '@tamanu/constants';
 import { isErrorUnknownAllow404s, useApi } from '../index';
 import { useVitalsSurveyQuery } from './useVitalsSurveyQuery';
-import { getConfigObject } from '../../utils';
-
-function hasHistoricalData(answer) {
-  if (!answer) return false;
-  const { records } = answer;
-  return Object.values(records).some(record => record.body);
-}
-
-function getSortedLogsByDate(logs) {
-  if (!logs) return [];
-  return logs.slice().sort((a, b) => b.date.localeCompare(a.date));
-}
-
-function getDatesAndRecords(data, surveyData, dateElementId) {
-  if (data.length > 0 === false || Boolean(surveyData) === false) {
-    return { recordedDates: [], records: [] };
-  }
-
-  const recordedDates = Object.keys(
-    data.find(vital => vital.dataElementId === dateElementId)
-      .records,
-  );
-
-  const elementIdToAnswer = data.reduce(
-    (dict, a) => ({ ...dict, [a.dataElementId]: a }),
-    {},
-  );
-
-  const records = surveyData.components
-    .filter(component => component.dataElementId !== dateElementId)
-    // Show current components or ones that have historical data in them
-    .filter(
-      component =>
-        component.visibilityStatus === VISIBILITY_STATUSES.CURRENT ||
-        hasHistoricalData(elementIdToAnswer[component.dataElementId]),
-    )
-    .map(component => {
-      const { id, config, validationCriteria, dataElement } = component;
-      const { records = {} } = elementIdToAnswer[dataElement.id] || {};
-      const configs = {
-        validationCriteria: getConfigObject(id, validationCriteria),
-        config: getConfigObject(id, config),
-      };
-      return recordedDates.reduce(
-        (state, date) => ({
-          ...state,
-          [date]: {
-            component,
-            recordedDate: date,
-            answerId: records[date]?.id,
-            value: records[date]?.body,
-            historyLogs: getSortedLogsByDate(records[date]?.logs),
-            ...configs,
-          },
-        }),
-        {
-          dataElementId: dataElement.id,
-          value: dataElement.name,
-          ...configs,
-        },
-      );
-    });
-
-  return { recordedDates, records };
-}
+import { getDatesAndRecords } from './useChartQuery';
 
 export const useVitals = encounterId => {
   const api = useApi();
